Replace old analysis script
